requirements.txt
requests==2.32.4
slack_bolt==1.23.0
python-dotenv==1.1.1